refactor: split tool formatting code
