+++ products/llm_analytics/backend/providers/formatters/tools_handler/models.py
@@ -1,0 +1,18 @@
+from dataclasses import dataclass
+from enum import Enum
+from typing import Any
+
+
+@dataclass
+class Tool:
+    """Internal representation of a tool/function"""
+
+    name: str
+    description: str
+    parameters: dict[str, Any]
+
+
+class ToolFormat(Enum):
+    OPENAI = "openai"
+    ANTHROPIC = "anthropic"
+    GEMINI = "gemini"
